+++ backend/app.py
@@ -1,4 +1,4 @@
-from flask import Flask
+from backend import Flask
 from models import db
 from config import Config
