+++ backend/email.py
@@ -1,5 +1,5 @@
 from threading import Thread
-from flask import current_app
+from backend import current_app
 from flask_mail import Message
 from app import mail
